requirements.txt
azure-functions
requests
azure.cosmos
azure-storage-queue